# sentinel/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Defentrax API
  description: |
    Control-plane REST API for Defentrax (security monitoring / lightweight SIEM).

    This specification documents **implemented** routes only (inventory from
    `sentinel/api/internal/server/server.go`). It does **not** invent endpoints.

    **Not exposed as REST (yet):**
    - No standalone `GET /api/v1/agents` — agents appear nested under `GET /api/v1/servers/{id}`.
    - No `/api/v1/metrics` — use `GET /api/v1/dashboard/stats`.

    **Page permissions:** panel routes require a `pages:*` key in addition to the
    data permission named on each operation. `pages:dashboard`, `pages:alerts`,
    `pages:alert_detail`, `pages:events`, `pages:servers`, `pages:server_detail`,
    `pages:rules`, `pages:notifications`, `pages:team`, `pages:roles`, `pages:audit`.
    `SUPER_ADMIN` and `ADMIN` bypass these checks. Plugins and API keys have no page key.

    Interactive UI: `GET /api/v1/docs` (Swagger UI). Raw spec: `GET /api/v1/openapi.yaml`.
  version: 0.2.0
  contact:
    name: Defentrax
servers:
  - url: http://localhost:8080
    description: Local default (API_PORT / PORT)
...
paths:
...
  /api/v1/saved-views:
    get:
      tags: [Views]
      operationId: listSavedViews
      summary: List saved alert or event filters for the current operator
      description: "Requires the read permission and page permission of the requested kind (`alerts` or `events`)."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      parameters:
        - name: kind
          in: query
          required: true
          schema:
            type: string
            enum: [alerts, events]
      responses:
        "200":
          description: Views
        "403":
          $ref: "#/components/responses/Error"
    post:
      tags: [Views]
      operationId: createSavedView
      summary: Save the current alert or event filter
      security:
        - SessionCookie: []
        - BearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedViewWrite"
      responses:
        "201":
          description: Created
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/saved-views/{id}:
    patch:
      tags: [Views]
      operationId: updateSavedView
      summary: Rename or replace a saved view
      security:
        - SessionCookie: []
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/PathUUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedViewWrite"
      responses:
        "200":
          description: Updated
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Views]
      operationId: deleteSavedView
      summary: Delete a saved view
      security:
        - SessionCookie: []
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/PathUUID"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sentinel_session
      description: "Server-side session cookie (name configurable via `SESSION_COOKIE_NAME`; default `sentinel_session`)."
    BearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: sent_
      description: "Personal API key (`Authorization: Bearer sent_…`)."
    BearerAgentToken:
      type: http
      scheme: bearer
      bearerFormat: sagt_
      description: "Agent token from enrollment (`Authorization: Bearer sagt_…`)."
  parameters:
    PathUUID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Standard error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          example: unauthorized
        message:
          type: string
        request_id:
          type: string
...
    SavedViewWrite:
      type: object
      required: [kind, name]
      properties:
        kind:
          type: string
          enum: [alerts, events]
        name:
          type: string
        is_default:
          type: boolean
        query:
          type: object
          additionalProperties:
            type: string
...